Fix circlular dependencies errors.
