+++ .gitignore
@@ -2,7 +2,7 @@ node_modules
 build
 .vscode
 *.d.ts
-!src/server/config/clowder.d.ts
+!src/common/clowder.d.ts
 !src/tests-visual-regression/global.d.ts
 dist
 public/*
