+++ src/common/consts.ts
@@ -1,1 +1,3 @@
 export const IS_PRODUCTION = process.env.NODE_ENV === 'production';
+export const IS_DEVELOPMENT = process.env.NODE_ENV === 'development';
+export const OPTIONS_HEADER_NAME = 'x-pdf-gen-options';
