+++ src/templates/automation-analytics/Components/ReportComponents/AutomationCalculator/TemplatesTable/Row.tsx
@@ -2,10 +2,11 @@ import React, { Fragment, FunctionComponent } from 'react';
 import { Tr, Td } from '../../../StyledPatternfly/Table';
 import { global_success_color_200 as globalSuccessColor200 } from '@patternfly/react-tokens';
 
-import { Template } from '../types';
+import { Template } from '../../types';
 import currencyFormatter from '../../../../Utilities/currencyFormatter';
-import ExpandedRowContents from './ExpandedRowContents';
-import { AutomationCalculatorExpandRowsComponentType } from '../../../ReportComponents/Standard/Components/types';
+import ExpandedRowContents, {
+  AutomationCalculatorExpandRowsComponentType,
+} from './ExpandedRowContents';
 import styled from 'styled-components';
 import * as Table from '@patternfly/react-table';
