+++ src/templates/automation-analytics/Components/ReportComponents/Standard/Components/types.ts
@@ -1,9 +1,6 @@
 import TemplateExplorerExpandedRow from './TemplateExplorerExpandedRow';
-import ExpandedRowContents from '../../AutomationCalculator/TemplatesTable/ExpandedRowContents';
 
 export type ExpandRowsComponentType = typeof TemplateExplorerExpandedRow;
-export type AutomationCalculatorExpandRowsComponentType =
-  typeof ExpandedRowContents;
 
 export enum ReportStandardExpandedRow {
   templateExplorer = 'templateExplorer',
