+++ src/templates/automation-analytics/Components/ReportComponents/types.ts
@@ -1,15 +1,45 @@
 import { FC } from 'react';
 import { ChartSchemaElement } from 'react-json-chart-builder';
 import { ApiReturnType } from '../ChartHelpers/types';
-import { ReportAutomationCalculatorProps } from './AutomationCalculator/types';
-import { ReportStandardProps } from './Standard/types';
+import { ReportStandardExpandedRow } from './Standard/Components/types';
+export * from './Standard/Components/types';
 
-export * from './AutomationCalculator/types';
-export * from './Standard/types';
+export type LegendEntry = Record<string, string | number>;
+export type TableHeaders = { key: string; value: string }[];
 
-export enum ReportLayout {
-  Standard = 'standard',
-  AutomationCalculator = 'automationCalculator',
+export interface ReportStandardProps extends BaseReportProps {
+  tableHeaders: TableHeaders;
+  expandedRowComponent?: ReportStandardExpandedRow;
+}
+
+export type ReportStandardDataType = ApiReturnType;
+
+export interface Template {
+  name: string;
+  id: number;
+  elapsed: number;
+  host_count: number;
+  total_count: number;
+  total_org_count: number;
+  total_cluster_count: number;
+  total_inventory_count: number;
+  successful_hosts_total: number;
+  successful_elapsed_total: number;
+  // Calculated fields
+  delta: number;
+  avgRunTime: number;
+  manualCost: number;
+  automatedCost: number;
+  enabled: boolean;
+  // Anything else accidentally having it
+  [key: string]: string | number | boolean;
+}
+
+export interface ReportAutomationCalculatorDataType
+  extends Omit<ApiReturnType, 'meta'> {
+  meta: {
+    legend: Template[];
+  };
 }
 
 export interface BaseReportProps {
@@ -19,6 +49,16 @@ export interface BaseReportProps {
   description: string;
   data: ApiReturnType;
   extraData: ApiReturnType;
+}
+
+export interface ReportAutomationCalculatorProps extends BaseReportProps {
+  tableHeaders: TableHeaders;
+  expandedRowComponent?: ReportStandardExpandedRow;
+}
+
+export enum ReportLayout {
+  Standard = 'standard',
+  AutomationCalculator = 'automationCalculator',
 }
 
 export type ReportComponentType =
